+++ src/main/java/org/opensearch/neuralsearch/processor/TextImageEmbeddingProcessor.java
@@ -177,7 +177,7 @@ public class TextImageEmbeddingProcessor extends AbstractProcessor {
             fieldMap,
             1,
             ProcessorDocumentUtils.getMaxDepth(sourceAndMetadataMap, clusterService, environment),
-            true
+            false
         );
     }